refactor action buttons
