+++ Halcyon.Web/src/features/profile/profile/delete-account-button.tsx
@@ -1,3 +1,5 @@
+import { useNavigate } from 'react-router';
+import { toast } from 'sonner';
 import {
     AlertDialog,
     AlertDialogAction,
@@ -9,27 +11,50 @@ import {
     AlertDialogTitle,
     AlertDialogTrigger,
 } from '@/components/ui/alert-dialog';
+import { useAuth } from '@/components/auth-provider';
 import { LoadingButton } from '@/components/loading-button';
+import { useDeleteAccount } from '@/features/profile/hooks/use-delete-account';
+import type { GetProfileResponse } from '@/features/profile/profile-types';
 
 type DeleteAccountButtonProps = {
-    onClick: () => void;
+    profile: GetProfileResponse;
     disabled?: boolean;
-    loading?: boolean;
     className?: string;
 };
 
 export function DeleteAccountButton({
-    onClick,
+    profile,
     disabled,
-    loading,
     className,
 }: DeleteAccountButtonProps) {
+    const navigate = useNavigate();
+
+    const { clearAuth } = useAuth();
+
+    const { mutate: deleteAccount, isPending: isDeleting } = useDeleteAccount();
+
+        function onDelete() {
+        deleteAccount(
+            {
+                version: profile.version,
+            },
+            {
+                onSuccess: () => {
+                    toast.success('Your account has been deleted.');
+                    clearAuth();
+                    navigate('/');
+                },
+                onError: (error) => toast.error(error.message),
+            }
+        );
+    }
+
     return (
         <AlertDialog>
             <AlertDialogTrigger asChild>
                 <LoadingButton
                     variant="destructive"
-                    loading={loading}
+                    loading={isDeleting}
                     disabled={disabled}
                     className={className}
                 >
@@ -48,8 +73,8 @@ export function DeleteAccountButton({
                 <AlertDialogFooter>
                     <AlertDialogCancel>Cancel</AlertDialogCancel>
                     <AlertDialogAction
-                        disabled={disabled || loading}
-                        onClick={onClick}
+                        disabled={disabled || isDeleting}
+                        onClick={onDelete}
                     >
                         Continue
                     </AlertDialogAction>
